% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{qcflow_update_registered_model}
\alias{qcflow_update_registered_model}
\title{Update a registered model}
\usage{
qcflow_update_registered_model(name, description, client = NULL)
}
\arguments{
\item{name}{The name of the registered model.}

\item{description}{The updated description for this registered model.}

\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}
}
\description{
Updates a model in the Model Registry.
}
