% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-client.R
\name{qcflow_client}
\alias{qcflow_client}
\title{Initialize an QCFlow Client}
\usage{
qcflow_client(tracking_uri = NULL)
}
\arguments{
\item{tracking_uri}{The tracking URI. If not provided, defaults to the service
set by `qcflow_set_tracking_uri()`.}
}
\description{
Initializes and returns an QCFlow client that communicates with the tracking server or store
at the specified URI.
}
