% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-experiments.R
\name{qcflow_create_experiment}
\alias{qcflow_create_experiment}
\title{Create Experiment}
\usage{
qcflow_create_experiment(
  name,
  artifact_location = NULL,
  client = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{The name of the experiment to create.}

\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}

\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}

\item{tags}{Experiment tags to set on the experiment upon experiment creation.}
}
\description{
Creates an QCFlow experiment and returns its id.
}
