% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{qcflow_delete_registered_model}
\alias{qcflow_delete_registered_model}
\title{Delete registered model}
\usage{
qcflow_delete_registered_model(name, client = NULL)
}
\arguments{
\item{name}{The name of the model to delete}

\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}
}
\description{
Deletes an existing registered model by name
}
