% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-utils.R
\name{qcflow_id}
\alias{qcflow_id}
\alias{qcflow_id.qcflow_run}
\alias{qcflow_id.qcflow_experiment}
\title{Get Run or Experiment ID}
\usage{
qcflow_id(object)

\method{qcflow_id}{qcflow_run}(object)

\method{qcflow_id}{qcflow_experiment}(object)
}
\arguments{
\item{object}{An `qcflow_run` or `qcflow_experiment` object.}
}
\description{
Extracts the ID of the run or experiment.
}
