% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{qcflow_log_metric}
\alias{qcflow_log_metric}
\title{Log Metric}
\usage{
qcflow_log_metric(
  key,
  value,
  timestamp = NULL,
  step = NULL,
  run_id = NULL,
  client = NULL
)
}
\arguments{
\item{key}{Name of the metric.}

\item{value}{Float value for the metric being logged.}

\item{timestamp}{Timestamp at which to log the metric. Timestamp is rounded to the nearest
integer. If unspecified, the number of milliseconds since the Unix epoch is used.}

\item{step}{Step at which to log the metric. Step is rounded to the nearest integer. If
unspecified, the default value of zero is used.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}
}
\description{
Logs a metric for a run. Metrics key-value pair that records a single float measure.
  During a single execution of a run, a particular metric can be logged several times.
  The QCFlow Backend keeps track of historical metric values along two axes: timestamp and step.
}
