% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{qcflow_predict}
\alias{qcflow_predict}
\title{Generate Prediction with QCFlow Model}
\usage{
qcflow_predict(model, data, ...)
}
\arguments{
\item{model}{The loaded QCFlow model flavor.}

\item{data}{A data frame to perform scoring.}

\item{...}{Optional additional arguments passed to underlying predict
methods.}
}
\description{
Performs prediction over a model loaded using
\code{qcflow_load_model()}, to be used by package authors
to extend the supported QCFlow models.
}
