% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{qcflow_rename_registered_model}
\alias{qcflow_rename_registered_model}
\title{Rename a registered model}
\usage{
qcflow_rename_registered_model(name, new_name, client = NULL)
}
\arguments{
\item{name}{The current name of the model.}

\item{new_name}{The new name for the model.}

\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}
}
\description{
Renames a model in the Model Registry.
}
