% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-runs.R
\name{qcflow_set_tag}
\alias{qcflow_set_tag}
\title{Set Tag}
\usage{
qcflow_set_tag(key, value, run_id = NULL, client = NULL)
}
\arguments{
\item{key}{Name of the tag. Maximum size is 255 bytes. This field is required.}

\item{value}{String value of the tag being logged. Maximum size is 500 bytes. This field is required.}

\item{run_id}{Run ID.}

\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}
}
\description{
Sets a tag on a run. Tags are run metadata that can be updated during a run and
 after a run completes.
}
