% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-globals.R
\name{qcflow_set_tracking_uri}
\alias{qcflow_set_tracking_uri}
\title{Set Remote Tracking URI}
\usage{
qcflow_set_tracking_uri(uri)
}
\arguments{
\item{uri}{The URI to the remote QCFlow server.}
}
\description{
Specifies the URI to the remote QCFlow server that will be used
to track experiments.
}
