% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking-ui.R
\name{qcflow_ui}
\alias{qcflow_ui}
\title{Run QCFlow User Interface}
\usage{
qcflow_ui(client, ...)
}
\arguments{
\item{client}{(Optional) An QCFlow client object returned from \link[qcflow]{qcflow_client}.
If specified, QCFlow will use the tracking server associated with the passed-in client. If
unspecified (the common case),
QCFlow will use the tracking server associated with the current tracking URI.}

\item{...}{Optional arguments passed to `qcflow_server()` when `x` is a path to a file store.}
}
\description{
Launches the QCFlow user interface.
}
\examples{
\dontrun{
library(qcflow)

# launch qcflow ui locally
qcflow_ui()

# launch qcflow ui for existing qcflow server
qcflow_set_tracking_uri("http://tracking-server:5000")
qcflow_ui()
}

}
